+++ __docs__/wonder-blocks-dropdown/multi-select-testing-snapshots.stories.tsx
@@ -4,6 +4,13 @@ import type {Meta, StoryObj} from "@storybook/react-vite";
 import {MultiSelect, OptionItem} from "@khanacademy/wonder-blocks-dropdown";
 import {themeModes} from "../../.storybook/modes";
 import {defaultPseudoStates, StateSheet} from "../components/state-sheet";
+import {
+    longText,
+    longTextWithNoWordBreak,
+} from "../components/text-for-testing";
+import {ScenariosLayout} from "../components/scenarios-layout";
+import {PropsFor} from "@khanacademy/wonder-blocks-core";
+import {sizing} from "@khanacademy/wonder-blocks-tokens";
 
 /**
  * The following stories are used to generate the pseudo states for the
@@ -41,6 +48,10 @@ const columns = [
         props: {disabled: true},
     },
     {
+        name: "Readonly",
+        props: {readOnly: true},
+    },
+    {
         name: "Error",
         props: {error: true},
     },
@@ -59,6 +70,10 @@ const columns = [
     {
         name: "With selection (All selected) + Disabled",
         props: {selectedValues: ["1", "2", "3"], disabled: true},
+    },
+    {
+        name: "With selection (All selected) + Readonly",
+        props: {selectedValues: ["1", "2", "3"], readOnly: true},
     },
 ];
 
@@ -79,6 +94,101 @@ export const StateSheetStory: Story = {
         );
     },
     parameters: {
-        pseudo: defaultPseudoStates,
+        pseudo: {
+            ...defaultPseudoStates,
+            // Using the focus selector instead so that the focus outline is shown
+            // when the select opener is clicked
+            focus: [...defaultPseudoStates.focusVisible],
+        },
+    },
+};
+
+const ControlledMultiSelect = (props: PropsFor<typeof MultiSelect>) => {
+    const [selectedValues, setSelectedValues] = React.useState(
+        props.selectedValues,
+    );
+    return (
+        <MultiSelect
+            {...props}
+            selectedValues={selectedValues}
+            onChange={setSelectedValues}
+        />
+    );
+};
+
+export const Scenarios: Story = {
+    render: () => {
+        const scenarios = [
+            {
+                name: "Long option item label",
+                props: {
+                    children: [
+                        <OptionItem label={longText} value="1" key="1" />,
+                    ],
+                    opened: true,
+                    style: {paddingBlockEnd: sizing.size_560},
+                },
+            },
+            {
+                name: "Long option item label with no word break",
+                props: {
+                    children: [
+                        <OptionItem
+                            label={longTextWithNoWordBreak}
+                            value="1"
+                            key="1"
+                        />,
+                    ],
+                    opened: true,
+                    style: {marginBlockEnd: sizing.size_560},
+                },
+            },
+            {
+                name: "Long selected option label",
+                props: {
+                    children: [
+                        <OptionItem label={longText} value="1" key="1" />,
+                        <OptionItem label={longText} value="2" key="2" />,
+                        <OptionItem label={longText} value="3" key="3" />,
+                    ],
+                    selectedValues: ["1"],
+                },
+            },
+            {
+                name: "Long selected option label with no word break",
+                props: {
+                    children: [
+                        <OptionItem
+                            label={longTextWithNoWordBreak}
+                            value="1"
+                            key="1"
+                        />,
+                        <OptionItem
+                            label={longTextWithNoWordBreak}
+                            value="2"
+                            key="2"
+                        />,
+                        <OptionItem
+                            label={longTextWithNoWordBreak}
+                            value="3"
+                            key="3"
+                        />,
+                    ],
+                    selectedValues: ["1"],
+                },
+            },
+        ];
+        return (
+            <ScenariosLayout scenarios={scenarios}>
+                {(props, name) => (
+                    <ControlledMultiSelect {...props} aria-label={name} />
+                )}
+            </ScenariosLayout>
+        );
+    },
+    globals: {
+        viewport: {
+            value: "small",
+        },
     },
 };