+++ __docs__/wonder-blocks-modal/modal-footer.stories.tsx
@@ -108,6 +108,20 @@ export default {
             // one-pane-dialog.stories.tsx
             disableSnapshot: true,
         },
+        a11y: {
+            // TODO(WB-1834): Fix the a11y violations and remove this.
+            config: {
+                rules: [
+                    // Disabling a11y violation: "Scrollable region must have
+                    // keyboard access (scrollable-region-focusable)".
+                    // ModalContent's scrollOverflow element is not focusable.
+                    {
+                        id: "scrollable-region-focusable",
+                        enabled: false,
+                    },
+                ],
+            },
+        },
     },
     argTypes: {
         children: {