+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 4.1.28
+
+### Patch Changes
+
+- Updated dependencies [0530fb2]
+    - @khanacademy/wonder-blocks-dropdown@10.6.0
+
+## 4.1.27
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+- Updated dependencies [6d35175]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-dropdown@10.5.6
+    - @khanacademy/wonder-blocks-icon@5.3.5
+    - @khanacademy/wonder-blocks-layout@3.1.42
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
+## 4.1.26
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-dropdown@10.5.5
+
 ## 4.1.25
 
 ### Patch Changes