+++ packages/wonder-blocks-button/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-button
 
+## 11.2.10
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-clickable@8.0.5
+    - @khanacademy/wonder-blocks-icon@5.3.5
+    - @khanacademy/wonder-blocks-progress-spinner@3.1.42
+    - @khanacademy/wonder-blocks-styles@0.2.37
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
+## 11.2.9
+
+### Patch Changes
+
+- 0fd41cc: Replace style prop with styles object prop for ActivityIconButton
+
 ## 11.2.8
 
 ### Patch Changes