+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 8.0.5
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+
 ## 8.0.4
 
 ### Patch Changes