+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 11.0.1
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-clickable@8.0.5
+    - @khanacademy/wonder-blocks-icon@5.3.5
+    - @khanacademy/wonder-blocks-styles@0.2.37
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
+## 11.0.0
+
+### Major Changes
+
+- 0fd41cc: Replace style prop with styles object prop for ActivityIconButton
+
+### Patch Changes
+
+- d36860e: Changes `ActivityIconButton` styling to allow longer labels in the same line.
+
 ## 10.5.7
 
 ### Patch Changes