+++ packages/wonder-blocks-labeled-field/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-labeled-field
 
+## 4.0.12
+
+### Patch Changes
+
+- 0530fb2: LabeledField: Fix passing down readOnly prop so it is a boolean
+
+## 4.0.11
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-layout@3.1.42
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
 ## 4.0.10
 
 ### Patch Changes