+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-link
 
+## 10.0.6
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-clickable@8.0.5
+    - @khanacademy/wonder-blocks-icon@5.3.5
+    - @khanacademy/wonder-blocks-styles@0.2.37
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
 ## 10.0.5
 
 ### Patch Changes