+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @khanacademy/wonder-blocks-popover
 
+## 6.1.47
+
+### Patch Changes
+
+- 5db127a: Fix issue where Popover focus management was interfering with Tabs focus management
+
+## 6.1.46
+
+### Patch Changes
+
+- Updated dependencies [70d6c08]
+    - @khanacademy/wonder-blocks-tokens@14.1.3
+    - @khanacademy/wonder-blocks-icon-button@11.0.1
+    - @khanacademy/wonder-blocks-modal@8.5.9
+    - @khanacademy/wonder-blocks-styles@0.2.37
+    - @khanacademy/wonder-blocks-tooltip@4.1.60
+    - @khanacademy/wonder-blocks-typography@4.2.27
+
+## 6.1.45
+
+### Patch Changes
+
+- Updated dependencies [0fd41cc]
+- Updated dependencies [d36860e]
+    - @khanacademy/wonder-blocks-icon-button@11.0.0
+    - @khanacademy/wonder-blocks-modal@8.5.8
+    - @khanacademy/wonder-blocks-tooltip@4.1.59
+
 ## 6.1.44
 
 ### Patch Changes